requirements.txt
aiogram
python-dotenv
requests
SQLAlchemy
gino
asyncpg
aiohttp
